+++ test/UT/mem_controller/test_ubse_mem_agent_task_manager.cpp
@@ -34,7 +34,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddFdExportTask)
 {
     UbseMemFdBorrowExportObj obj{};
     obj.req.name = "test_fd_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddFdExportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR, UbseMemAgentTaskManager::AddFdExportTask(taskId, obj)); // Should fail as task already exists
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::DeleteTaskObj(taskId));
@@ -44,7 +44,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteFdExportTask)
 {
     UbseMemFdBorrowExportObj obj{};
     obj.req.name = "test_fd_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddFdExportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteFdExportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteFdExportTask(taskId)); // Should fail as task does not exist
@@ -54,7 +54,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetFdExportTaskObj)
 {
     UbseMemFdBorrowExportObj obj{};
     obj.req.name = "test_fd_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddFdExportTask(taskId, obj);
     UbseMemFdBorrowExportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetFdExportTaskObj("test_fd_export", "", returnObj));
@@ -66,7 +66,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddNumaExportTask)
 {
     UbseMemNumaBorrowExportObj obj{};
     obj.req.name = "test_numa_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddNumaExportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR,
               UbseMemAgentTaskManager::AddNumaExportTask(taskId, obj)); // Should fail as task already exists
@@ -77,7 +77,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteNumaExportTask)
 {
     UbseMemNumaBorrowExportObj obj{};
     obj.req.name = "test_numa_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddNumaExportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteNumaExportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteNumaExportTask(taskId)); // Should fail as task does not exist
@@ -87,7 +87,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetNumaExportTaskObj)
 {
     UbseMemNumaBorrowExportObj obj{};
     obj.req.name = "test_numa_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddNumaExportTask(taskId, obj);
     UbseMemNumaBorrowExportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetNumaExportTaskObj("test_numa_export", "", returnObj));
@@ -99,7 +99,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddShareExportTask)
 {
     UbseMemShareBorrowExportObj obj{};
     obj.req.name = "test_share_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddShareExportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR,
               UbseMemAgentTaskManager::AddShareExportTask(taskId, obj)); // Should fail as task already exists
@@ -110,7 +110,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteShareExportTask)
 {
     UbseMemShareBorrowExportObj obj{};
     obj.req.name = "test_share_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddShareExportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteShareExportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteShareExportTask(taskId)); // Should fail as task does not exist
@@ -120,7 +120,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetShareExportTaskObj)
 {
     UbseMemShareBorrowExportObj obj{};
     obj.req.name = "test_share_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddShareExportTask(taskId, obj);
     UbseMemShareBorrowExportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetShareExportTaskObj("test_share_export", returnObj));
@@ -132,7 +132,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddAddrExportTask)
 {
     UbseMemAddrBorrowExportObj obj{};
     obj.req.name = "test_addr_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddAddrExportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR,
               UbseMemAgentTaskManager::AddAddrExportTask(taskId, obj)); // Should fail as task already exists
@@ -143,7 +143,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteAddrExportTask)
 {
     UbseMemAddrBorrowExportObj obj{};
     obj.req.name = "test_addr_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddAddrExportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteAddrExportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteAddrExportTask(taskId)); // Should fail as task does not exist
@@ -153,7 +153,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetAddrExportTaskObj)
 {
     UbseMemAddrBorrowExportObj obj{};
     obj.req.name = "test_addr_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddAddrExportTask(taskId, obj);
     UbseMemAddrBorrowExportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetAddrExportTaskObj("test_addr_export", "", returnObj));
@@ -165,7 +165,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddFdImportTask)
 {
     UbseMemFdBorrowImportObj obj{};
     obj.req.name = "test_fd_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddFdImportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR, UbseMemAgentTaskManager::AddFdImportTask(taskId, obj)); // Should fail as task already exists
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::DeleteTaskObj(taskId));
@@ -175,7 +175,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteFdImportTask)
 {
     UbseMemFdBorrowImportObj obj{};
     obj.req.name = "test_fd_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddFdImportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteFdImportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteFdImportTask(taskId)); // Should fail as task does not exist
@@ -185,7 +185,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetFdImportTaskObj)
 {
     UbseMemFdBorrowImportObj obj{};
     obj.req.name = "test_fd_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddFdImportTask(taskId, obj);
     UbseMemFdBorrowImportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetFdImportTaskObj("test_fd_import", returnObj));
@@ -197,7 +197,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddNumaImportTask)
 {
     UbseMemNumaBorrowImportObj obj{};
     obj.req.name = "test_numa_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddNumaImportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR,
               UbseMemAgentTaskManager::AddNumaImportTask(taskId, obj)); // Should fail as task already exists
@@ -208,7 +208,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteNumaImportTask)
 {
     UbseMemNumaBorrowImportObj obj{};
     obj.req.name = "test_numa_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddNumaImportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteNumaImportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteNumaImportTask(taskId)); // Should fail as task does not exist
@@ -218,7 +218,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetNumaImportTaskObj)
 {
     UbseMemNumaBorrowImportObj obj{};
     obj.req.name = "test_numa_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddNumaImportTask(taskId, obj);
     UbseMemNumaBorrowImportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetNumaImportTaskObj("test_numa_import", returnObj));
@@ -230,7 +230,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddShareImportTask)
 {
     UbseMemShareBorrowImportObj obj{};
     obj.req.name = "test_share_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddShareImportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR,
               UbseMemAgentTaskManager::AddShareImportTask(taskId, obj)); // Should fail as task already exists
@@ -241,7 +241,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteShareImportTask)
 {
     UbseMemShareBorrowImportObj obj{};
     obj.req.name = "test_share_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddShareImportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteShareImportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteShareImportTask(taskId)); // Should fail as task does not exist
@@ -251,7 +251,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetShareImportTaskObj)
 {
     UbseMemShareBorrowImportObj obj{};
     obj.req.name = "test_share_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddShareImportTask(taskId, obj);
     UbseMemShareBorrowImportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetShareImportTaskObj("test_share_import", returnObj));
@@ -263,7 +263,7 @@ TEST_F(TestUbseMemAgentTaskManager, AddAddrImportTask)
 {
     UbseMemAddrBorrowImportObj obj{};
     obj.req.name = "test_addr_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::AddAddrImportTask(taskId, obj));
     EXPECT_EQ(UBSE_ERROR,
               UbseMemAgentTaskManager::AddAddrImportTask(taskId, obj)); // Should fail as task already exists
@@ -274,7 +274,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteAddrImportTask)
 {
     UbseMemAddrBorrowImportObj obj{};
     obj.req.name = "test_addr_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddAddrImportTask(taskId, obj);
     EXPECT_TRUE(UbseMemAgentTaskManager::DeleteAddrImportTask(taskId));
     EXPECT_FALSE(UbseMemAgentTaskManager::DeleteAddrImportTask(taskId)); // Should fail as task does not exist
@@ -284,7 +284,7 @@ TEST_F(TestUbseMemAgentTaskManager, GetAddrImportTaskObj)
 {
     UbseMemAddrBorrowImportObj obj{};
     obj.req.name = "test_addr_import";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddAddrImportTask(taskId, obj);
     UbseMemAddrBorrowImportObj returnObj{};
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::GetAddrImportTaskObj("test_addr_import", returnObj));
@@ -296,7 +296,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteTaskObj)
 {
     UbseMemFdBorrowExportObj obj{};
     obj.req.name = "test_fd_export";
-    uint64_t taskId = UbseMemAgentTaskManager::GetTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GetTaskId();
     UbseMemAgentTaskManager::AddFdExportTask(taskId, obj);
     EXPECT_EQ(UBSE_OK, UbseMemAgentTaskManager::DeleteTaskObj(taskId));
 }
@@ -304,7 +304,7 @@ TEST_F(TestUbseMemAgentTaskManager, DeleteTaskObj)
 TEST_F(TestUbseMemAgentTaskManager, GetTaskId)
 {
     EXPECT_EQ(0, UbseMemAgentTaskManager::GetTaskId());
-    uint64_t taskId = UbseMemAgentTaskManager::GenerateTaskId();
+    uint32_t taskId = UbseMemAgentTaskManager::GenerateTaskId();
     EXPECT_EQ(taskId + 1, UbseMemAgentTaskManager::GetTaskId());
 }
